sinks/redis/requirements.txt
redis 
kafka-python